Fix SSH, simplify configuration.
Configuration is now all based on a docker image naming convention.
There's no need to invoke the contain to fetch details. They are going
to be specified on the command line, or in a defaults file in the host
user's home directory. If the configuration file is lost, listings of
available tags can be derived by filtering `docker rmi`.

Fixed SSH agent forwarding when running as other than `root`.

Updated `README.md` to reflect the new image naming convention.