Pipe formulas out of Docker images.
Redefined formula as bundles so they would be easier to tar and
transfer, plus they can become more complicated, including any resources
they might need for configuration.

Now able to specify a formula in a Docker image using `docker:///` and
if that image is a Linux with `bash` and `tar` installed, the formula
will be extracted by running the image from a `bash` entry point,
extracting the formula using `tar`. This assumes a `bash` at `/bin/bash`.

Next is to create a bootstrapped environment, and to also install
homeport so it can be used in any homeport image.