Expose Docker inside Homeport, specify home dir.
More than enough Docker inside homeport, including mappings to your home
directory, but not to other operating system directories. That is, if
you map a volume in your home directory in Homeport, it will reference
the same directory in the host machine.

Also in this commit, the ability to specify a home directory to `docker
run` using the `--home` switch.

Closes #80.
Closes #84.