Rename `uid` and `gid` on boot.
Instead of using bindfs to map uid and gid, actually remap the uid and
gid based on the uid and gid in the user's home directory.

This will present a problem if we're ever trying to use resources that
have been given assigned to the `homeport` user on the drive. Which, as
I'm committing this, I'm starting to wonder if I shouldn't fall back to
running as `root`, but I'm sure that has it's own issues.

See: https://github.com/docker/docker/issues/7198

All this is going going to lead to a caveat of some sort, some sort of
entry in the FAQ. It is either, why are permissions wrong or why is it
slow? Slow is not going to have a simple explaination. Permissions do
have a simple explaination, you need kernel support for mapping on Linux.

We could also move to a convention, that permissions are alreays `501`
and `20` and that would make it work for a lot of people. You'd have to
move dialout, or learn to live with it, but it would be consistent for
anyone who is the primary user on their Mac. It falls down then they go
a remote machine they do not control.

If I did want to use a service remap, the best bet would be to piggyback
on an NFS mounting service on OS X, run NFS in the container on Linux.
NFSv4 has uid mapping.