SSH configuration Homeport.
Most of what the homeport wrapper scripts do is generate an SSH
configuration on the fly and then execute a command. If we simply
generate an SSH config and place it on the users `~/.ssh`, we don't have
to launch these wrappers anymore.

And when we don't have to do all that work, we can probably get
everything to fit into a single script so installation can be as simple
as running homeport, and the default action is to cat the script to
standard out. That or simply curl homeport from GitHub.