Start of single file Homeport.
Life is easier now that Docker Machine is a thing of the past. Got
`docker run` working in the single `homeport`. Seems that all that
remains is a means to build a homeport image, because we can always just
download the `homeport` script.

We can assume Docker too, so we can also cat the script or copy it from
an image.

All that remains is a way to build the image, though, and we can get rid
of the rest.